requirements.txt
Django==1.7.2
python-mpd2==0.5.4
